+++ .github/workflows/ccpp.yml
@@ -22,7 +22,7 @@ jobs:
         cd ..
     - name: Install other dependencies
       run: |
-        sudo apt-get install -y libdwarf-dev libelf-dev libsqlite3-dev gperf nodejs npm bison flex
+        sudo apt-get install -y libdwarf-dev libelf-dev libsqlite3-dev gperf nodejs npm bison flex libbpfcc-dev
         sudo npm install vsce -g
     - name: Make everything
       run: make
